Slight re-org of some code; establish global tracer when factory creates the OpenTelemetry instance for the service registry
