+++ bom/pom.xml
@@ -802,6 +802,11 @@
                 <artifactId>helidon-telemetry-opentelemetry-config</artifactId>
                 <version>${helidon.version}</version>
             </dependency>
+            <dependency>
+                <groupId>io.helidon.webserver.observe</groupId>
+                <artifactId>helidon-webserver-observe-telemetry-tracing</artifactId>
+                <version>${helidon.version}</version>
+            </dependency>
 
             <!-- tracing -->
             <dependency>
