+++ webserver/observe/pom.xml
@@ -37,5 +37,6 @@
         <module>metrics</module>
         <module>log</module>
         <module>tracing</module>
+        <module>telemetry</module>
     </modules>
 </project>
